+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * The matrix interface for fitting `fit_xy()` now works for censored regression models.
 
-* Predictions of survival probabilities for proportional hazards models with the `"survival"` and `"glmnet"` engines are now calculated by `summary.survfit()` (#221).
+* Predictions of survival probabilities are now calculated via `summary.survfit()` for `proportional_hazards()` models with the `"survival"` and `"glmnet"` engines, `bag_tree()` models with the `"rpart"` engine, `decision_tree()` models with the `"partykit"` engines, as well as `rand_forest()` models with the `"partykit"` engine (#221, #224). 
 
 * Added internal `survfit_summary_*()` helper functions (#216).